Add pin page logic in PVDMA device model.
In order to trigger the VFIO_IOMMU_MAP_DMA, we introduced
vfio_container_fd inside PCIDevice, to keep record the value
of the container fd.

We introduced dma_bitmap, which is indexed by bdf, to tell the
VM if the corresponding PCI device is an emulated one. If the
answer is yes, there's no need to pin the DMA pages.

Signed-off-by: Yu Zhang <yu.c.zhang@linux.intel.com>
Signed-off-by: Luwei Kang <luwei.kang@intel.com>