Introduce an emulated device for PVDMA.
An emulated PCI device is introduced to expose a MMIO
range, which can serve as the interface between the
front end PV DMA driver and the qemu side.

Note: Once the emulated PCI device is exposed, qemu will
not pin the guest memory - this is only an expediency to
solve the pre-pin issue.

Signed-off-by: Yu Zhang <yu.c.zhang@linux.intel.com>
Signed-off-by: Luwei Kang <luwei.kang@intel.com>